requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
crayons==0.3.0
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.11.0
pyparsing==2.4.4
pytest==7.4.4
requests==2.22.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.6
wcwidth==0.1.7
webdriver-manager==2.2.0
xlrd==1.2.0
zipp==0.6.0
